Add IDs to sections without and fix misc typos
This commit adds some IDs to blogposts missing them, so that relative
links to sections of the page will work. This includes an ID and link in
the sticky template post, so that I will remember to include them in
future posts.

It also fixes a misc typo, changing 'afterimge' to 'afterimage', and
makes the first sentence of the 'we built a PC' post slightly less
abrupt.